Moved groovyProjects.gradle.kts into buildSrc and renamed to ConfigureJavaProjects
