+++ build.gradle
@@ -208,7 +208,7 @@ subprojects {
     version = rootProject.version
 
     if (project in groovyProjects) {
-        apply from: "$rootDir/gradle/groovyProject.gradle.kts"
+        apply plugin: "configure-java-projects"
         apply from: "$rootDir/gradle/testWithUnknownOS.gradle"
         apply from: "$rootDir/gradle/java9plus.gradle"
         check.dependsOn ":docs:checkstyleApi"
